pyproject.toml
[project]
name = "mo-bills"
version = "0.1.0"
description = "Scraper for Missouri House of Representatives bills"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.28.1",
    "playwright>=1.40.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "langchain>=0.3.17",
    "langchain-openai>=0.3.2",
    "langchain-community>=0.3.17",
    "langgraph>=0.2.59",
    "pypdf>=5.1.0",
    "tiktoken>=0.8.0",
]

[project.scripts]
scrape-mo-bills = "ingestion.bills.scrape_mo_house_bills:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]

[dependency-groups]
dev = []
